feat: entity, dto 객체 멤버변수 타입 변환 - entity의 아이디 타입을 Long으로 변환 - entity,dto 멤버변수 타입을 wrapper class로 변환

+++ src/main/java/com/addiction/alertHistory/repository/AlertHistoryRepository.java
@@ -11,17 +11,17 @@ import com.addiction.alertHistory.entity.AlertHistory;
 
 public interface AlertHistoryRepository {
 
-	void deleteByUserId(int userId);
+	void deleteByUserId(Long userId);
 
-	List<AlertHistory> findByUserId(int userId);
+	List<AlertHistory> findByUserId(Long userId);
 
-	Page<AlertHistory> findByUserId(int userId, Pageable pageable);
+	Page<AlertHistory> findByUserId(Long userId, Pageable pageable);
 
 	Optional<AlertHistory> findById(Long id);
 
-	boolean hasUncheckedAlerts(int userId);
+	boolean hasUncheckedAlerts(Long userId);
 
-	boolean hasFriendCode(int userId, String friendCode);
+	boolean hasFriendCode(Long userId, String friendCode);
 
 	AlertHistory save(AlertHistory alertHistory);
